add simple example

+++ examples/branchless.rs
@@ -3,6 +3,7 @@ use std::num::Wrapping;
 
 pub fn main() {
     let args = parse_args();
+
     let (v1, v2, bools) = fill_vecs(&args);
 
     let a = do_branchless_work(&v1, &v2, &bools);
@@ -13,13 +14,13 @@ pub fn main() {
 }
 
 #[inline(never)]
-fn do_branchless_work(v1: &[u32], v2: &[u32], bools: &[bool]) -> Wrapping<u128> {
-    let mut a: Wrapping<u128> = Wrapping(0);
+fn do_branchless_work(v1: &[u32], v2: &[u32], bools: &[bool]) -> Wrapping<u32> {
+    let mut a: Wrapping<u32> = Wrapping(0);
 
     for _ in 0..100 {
         for j in 0..v1.len() {
             let vals = [v1[j], v2[j]];
-            a += vals[bools[j] as usize] as u128;
+            a += vals[bools[j] as usize];
         }
     }
